Improvements
One form to create/update employees. Removed the /create-user route and form and replaced it with a /manage-user route and form.
Employees are created on the /manage-employee route. Employees are updated on the /manage-user/${userId} route.
Behind the scenes, mongoose is only ever performing a User.update function, never a User.create. It makes use of the upsert function to create a new employee document if it doesn't find a match by authId.
react-hook-form makes use of setting the default values asynchronously if on the /manage-user/${userId} dynamic route.
Typescript has been complaining a lot and I've silenced it the best I can.